Split Vite decorator plugin entrypoint
